Move subscription to genome browser location change into a separate hook
