+++ src/content/app/genome-browser/components/browser-image/BrowserImage.tsx
@@ -15,23 +15,11 @@
  */
 
 import React, { useRef, useEffect, memo } from 'react';
-import { useNavigate } from 'react-router-dom';
-import { useDispatch, useSelector } from 'react-redux';
+import { useSelector } from 'react-redux';
 import classNames from 'classnames';
-import {
-  BrowserCurrentLocationUpdateAction,
-  BrowserTargetLocationUpdateAction,
-  IncomingActionType
-} from '@ensembl/ensembl-genome-browser';
-
-import * as urlFor from 'src/shared/helpers/urlHelper';
-import {
-  buildFocusIdForUrl,
-  parseFocusObjectId
-} from 'src/shared/helpers/focusObjectHelpers';
-import { getChrLocationStr } from 'src/content/app/genome-browser/helpers/browserHelper';
 
 import useGenomeBrowser from 'src/content/app/genome-browser/hooks/useGenomeBrowser';
+import useGenomeBrowserPosition from 'src/content/app/genome-browser/hooks/useGenomeBrowserPosition';
 
 import BrowserCogList from '../browser-cog/BrowserCogList';
 import { ZmenuController } from 'src/content/app/genome-browser/components/zmenu';
@@ -41,17 +29,9 @@ import Overlay from 'src/shared/components/overlay/Overlay';
 import { BROWSER_CONTAINER_ID } from 'src/content/app/genome-browser/constants/browser-constants';
 
 import {
-  getBrowserActiveGenomeId,
-  getBrowserActiveFocusObjectId
-} from 'src/content/app/genome-browser/state/browser-general/browserGeneralSelectors';
-import {
   getRegionEditorActive,
   getRegionFieldActive
 } from 'src/content/app/genome-browser/state/browser-general/browserGeneralSelectors';
-import {
-  updateActualChrLocation,
-  ChrLocation
-} from 'src/content/app/genome-browser/state/browser-general/browserGeneralSlice';
 import { getBrowserNavOpenState } from 'src/content/app/genome-browser/state/browser-nav/browserNavSelectors';
 
 import styles from './BrowserImage.scss';
@@ -62,62 +42,12 @@ export const BrowserImage = () => {
   const { activateGenomeBrowser, clearGenomeBrowser, genomeBrowser } =
     useGenomeBrowser();
 
-  const activeGenomeId = useSelector(getBrowserActiveGenomeId);
-  const activeFocusId = useSelector(getBrowserActiveFocusObjectId);
+  useGenomeBrowserPosition();
+
   const isNavbarOpen = useSelector(getBrowserNavOpenState);
   const isRegionEditorActive = useSelector(getRegionEditorActive);
   const isRegionFieldActive = useSelector(getRegionFieldActive);
   const isDisabled = isRegionEditorActive || isRegionFieldActive;
-
-  const dispatch = useDispatch();
-  const navigate = useNavigate();
-
-  useEffect(() => {
-    const positionUpdate = (
-      action:
-        | BrowserCurrentLocationUpdateAction
-        | BrowserTargetLocationUpdateAction
-    ) => {
-      if (action.type === IncomingActionType.CURRENT_POSITION) {
-        const { stick, start, end } = action.payload;
-        const chromosome = stick.split(':')[1];
-        dispatch(updateActualChrLocation([chromosome, start, end]));
-      } else if (action.type === IncomingActionType.TARGET_POSITION) {
-        const { stick, start, end } = action.payload;
-        const chromosome = stick.split(':')[1];
-        const chrLocation = [chromosome, start, end] as ChrLocation;
-
-        const newFocus = buildFocusIdForUrl(
-          parseFocusObjectId(activeFocusId as string)
-        );
-
-        navigate(
-          urlFor.browser({
-            genomeId: activeGenomeId,
-            focus: newFocus,
-            location: getChrLocationStr(chrLocation)
-          }),
-          {
-            replace: true
-          }
-        );
-      }
-    };
-
-    const subscriptionToActualPotitionMessages = genomeBrowser?.subscribe(
-      IncomingActionType.CURRENT_POSITION,
-      positionUpdate
-    );
-    const subscriptionToTargetPotitionMessages = genomeBrowser?.subscribe(
-      IncomingActionType.TARGET_POSITION,
-      positionUpdate
-    );
-
-    return () => {
-      subscriptionToActualPotitionMessages?.unsubscribe();
-      subscriptionToTargetPotitionMessages?.unsubscribe();
-    };
-  }, [genomeBrowser]);
 
   useEffect(() => {
     if (!genomeBrowser) {
